ssvep/requirements.txt
# SSVEP demo - standalone (BrainAccess MINI cap)
# Python 3.10+

brainaccess
matplotlib
numpy
PyYAML
pygame
scipy
